# PERF: DataFrame constructor