```
from asv_bench.benchmarks.frame_ctor import *

self = FromArrays()
self.setup()

%timeit self.time_frame_from_arrays_int()
2.77 ms ± 129 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # <- master
2.11 ms ± 57.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # <- PR
```